- type cleanups in places where pointers are used as unique identifiers - working implementation of FindCycles glue code - modified lua/Makefile to build a shared library (for glue code)
git-svn-id: https://semistable.com/svn/debgraph@63 d95f148e-e248-0410-ae72-c04dada67d05

+++ trunk/edge.cc
@@ -57,8 +57,8 @@ const Node *Edge::getToNode() {
 string Edge::toGraphviz() {
 	char source_gid[11], sink_gid[11];
 	// Use pointer value as unique id
-	sprintf(source_gid, "%u", (unsigned int)from);
-	sprintf(sink_gid, "%u", (unsigned int)to);
+	sprintf(source_gid, "%u", (size_t)from);
+	sprintf(sink_gid, "%u", (size_t)to);
 	string output("\t node");
 	output += source_gid + string(" -> node") + sink_gid;
 	output += " [ label = \"" + this->getTypeString() + "\" ];\n";
